+++ CHANGELOG.md
@@ -3,6 +3,9 @@ These are selected quick notes for developers. For all changes, see the
 
 8.x-3.3:
 
+* Configuration: sp_cert_folder has been removed. sp_x509_certificate and
+  sp_private_key can now hold values with a 'file:' prefix.
+
 * SamlService::$samlAuth was changed into an array. (This is not considered
   part of the interface. SamlService::getSamlAuth(), which should be used for
   getting this object, is still backward compatible.) Passing the new argument